+++ crates/dbg-swc/scripts/check-size.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+set -eu
+
+cargo run -- minify ensure-size --no-terser > list.txt
+cat list.txt | xargs -L 1 cargo run -- minify reduce